friendly_enigma/requirements.txt
django==4.0.6
djangorestframework==3.13.1
django-model-utils==4.2.0
django-linear-migrations
django-phonenumber-field[phonenumbers]
django-request-logging
drf-yasg
gunicorn==20.1.0
psycopg2==2.9.1
pyJWT
pytz==2021.1
whitenoise==5.3.0
Celery==5.2.2
Pillow==9.0.1
python-memcached==1.59
django-storages[azure]
model-bakery==1.3.3
django-model-utils== 4.2.0
django-multiselectfield==0.1.12
django-import-export==2.7.0
pymediainfo==5.1.0
channels-redis==3.3.1
channels_rabbitmq==4.0.0
channels>=3.0.3,<4.0
uvicorn[standard]
daphne
fcm-django==1.0.8
jsonfield==3.1.0
django-celery-beat==2.2.1
django-admin-async-upload==3.0.4
django-simple-history==3.0.0
pytest==7.0.1
